+++ CHANGELOG.md
@@ -19,8 +19,9 @@
   * Add `cardThreeDSecureDidCancel()` to `CardVaultDelegate`
   * Add `cardThreeDSecureWillLaunch()` to `CardVaultDelegate`
   * Add `cardThreeDSecureDidFinish()` to `CardVaultDelegate`
-* PayPalWebPayments
-  * Add `vault(url:)` method to `PayPalWebCheckoutClient`
+* PayPalWebPayments  
+  * Add `PayPalVaultRequest` type for interacting with `vault` method
+  * Add `vault(_:)` method to `PayPalWebCheckoutClient`
   * Add `PayPalVaultResult` type to return vault result
   * Add `PayPalVaultDelegate` to handle results from vault flow
   * Add `PayPalWebCheckoutClientError.paypalVaultResponseError` for missing or invalid response from vaulting